+++ src/Hybrid.h
@@ -82,8 +82,7 @@ private:
 
   // parameters from json for the solver
   int elementOrder;
-  uint8_t timeOrder;
-  // uint32_t
+  int timeOrder;
   int numSubsteps;
   std::string preconditioner;
   std::string solverType;
@@ -110,9 +109,9 @@ void Hybrid<S,A,I>::init(std::vector<HOVolumes<S>>& _euler) {
   std::cout << "Inside Hybrid::init with " << _euler.size() << " volumes" << std::endl;
 
   solver.set_re_d_(100.0);
-  solver.set_elemorder_d_(elementOrder);
-  solver.set_timeorder_d_(timeOrder);
-  solver.set_numsteps_d_(numSubsteps);
+  solver.set_elemorder_d_((uint8_t)elementOrder);
+  solver.set_timeorder_d_((uint8_t)timeOrder);
+  solver.set_numsteps_d_((uint32_t)numSubsteps);
 
   for (auto &coll : _euler) {
     // transform to current position
